Update changes view to better measure when a commit was created
To have more precise metrics it'll measure the creation of a commit as the first time it was available for review (PR). This way, we avoid changing our processes as a whole.

The issue is that when merging a PR, it creates a new commit that seems to be created just at the time of merge. To measure lead time, it's better to consider the time it took to review the PR. For that, it sets the time of the merge commit as the PR creation time.